third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/accumulation-per-property-002-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] math-depth: integer
  assert_equals: The value should be -3 at 0ms expected "-3" but got "-2"
[FAIL] position-visibility: "anchors-valid" onto "always"
  assert_equals: The value should be anchors-valid at 0ms expected "anchors-valid" but got "always"
[FAIL] stroke-dasharray: dasharray
  assert_equals: The value should be 1px, 2px, 3px, 4px, 5px at 0ms expected "1px, 2px, 3px, 4px, 5px" but got "7px, calc(30% + 2px), 5px, 10px, calc(30% + 5px), 3px, 8px, calc(30% + 3px), 6px, 11px, calc(30% + 1px), 4px, 9px, calc(30% + 4px), 7px"
[FAIL] text-decoration: "overline" onto "underline"
  assert_equals: The value should be overline at 0ms expected "overline" but got "overline rgb(0, 0, 0)"
[FAIL] text-decoration: "underline" onto "overline"
  assert_equals: The value should be underline at 0ms expected "underline" but got "underline rgb(0, 0, 0)"
[FAIL] scale with two unspecified values
  assert_equals: The value should be -2 at 0ms expected "-2" but got "-6"
[FAIL] scale with one unspecified value
  assert_equals: The value should be -2 at 0ms expected "-2" but got "-6"
[FAIL] scale
  assert_equals: The value should be 0 -1 -2 at 0ms expected "0 -1 -2" but got "-2 -4 -6"
Harness: the test ran to completion.
